requirements.txt
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
coverage==7.6.10
cycler==0.12.1
filelock==3.16.1
Flask==3.1.0
fonttools==4.55.3
fsspec==2024.10.0
graphviz==0.20.3
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.8
lightning-utilities==0.11.9
MarkupSafe==2.1.5
matplotlib==3.10.0
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.1
packaging==24.2
pillow==11.0.0
pluggy==1.5.0
pyparsing==3.2.0
pytest==8.3.4
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
requests==2.32.3
setuptools==75.6.0
six==1.17.0
sympy==1.13.1
torch==2.6.0.dev20241229+cu126
torchaudio==2.6.0.dev20241230+cu126
torchmetrics==1.6.1
torchsummary==1.5.1
torchvision==0.22.0.dev20241230+cu126
torchviz==0.0.3
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.45.1
